Continued updating codebook entries (core prep up to abrasion)
Most of these changes are moving information from inclusion/exclusion criteria down to examples, and removing redundant text. Identified a logical conflict in abrasion, where there was a note that abrasion had to be noted as having a role in preparing core platforms (even though the code was modified a while back to also include abrasion to sharpen edges and other kinds of shaping practices). That is now removed.