Minor change to hard hammer and anvil rework.
Similar to the issues with coding hard hammer percussion, use of an anvil was not represented well in the codebook. I reworked it to be closer to what I have in mind. Use of an anvil can be present partly independently of bipolar percussion, as the original Clarkson citation suggests. I made it clearer that there only needs to be evidence in figures if it is ambiguous in the text about whether bipolar refers to traditional bipolar percussion, or removals of flakes from two separate poles without an anvil (what I refer to as bidirectional).  Also added in scaled pieces and pièces esquillées as important examples. Anvils now encompass any hard substrate.