Updating how inclusion/exclusion criteria are handled
After spending some time messing with E5. I think its worthwhile breaking down the inclusion/exclusion criteria into more discrete units. inclusion and exclusion criteria for text, and inclusion/exclusion criteria for figures.

So, the goal is we never actually make a presence/absence decision. We just report which criteria are met, then automatically calculate it so we can reduce decision making.

Here's an example of how this is structured for now:

**Inclusion criteria text:**
Code if sufficient evidence in text
**Inclusion criteria figures:**
Code if sufficient evidence in figures
**Exclusion criteria text:**
Code if insufficient evidence in text. Code if described as absent.
**Exclusion criteria figures**
Code if insufficient evidence in figures.
**Inclusion logic** Hierarchical core reduction present AND (sufficient evidence in text OR sufficient evidence in figures)
**Exclusion logic** Hierarchical core reduction absent OR (sufficient evidence in text AND  sufficient evidence in figures)